remove unnecessary files and dependencies
this is for a good reason i swear to god. basically, msfvenom (and apktool) will not inject backdoor payloads into APKs over a certain file size (65565 bytes or something like that i don't know). this took way less time than i thought it would to figure out the solution, but with these dependencies the apk is too large for msfvenom to handle